refactors
- upgarde execa
- remove moment
- pull ffmpeg/execa logic out into main process
