+++ package.json
@@ -67,8 +67,8 @@
     "i18next-parser": "^7.6.0",
     "icon-gen": "^3.0.0",
     "ky": "^0.33.1",
+    "luxon": "^3.3.0",
     "mkdirp": "^1.0.3",
-    "moment": "^2.29.4",
     "mousetrap": "^1.6.5",
     "p-map": "^5.5.0",
     "patch-package": "^6.2.1",
@@ -103,7 +103,7 @@
     "electron-is-dev": "^2.0.0",
     "electron-store": "5.1.1",
     "electron-unhandled": "^4.0.1",
-    "execa": "^5.0.0",
+    "execa": "^7.1.1",
     "file-type": "16",
     "file-url": "^3.0.0",
     "fs-extra": "^8.1.0",
