+++ public/electron.js
@@ -10,6 +10,7 @@ const yargsParser = require('yargs-parser');
 const JSON5 = require('json5');
 const remote = require('@electron/remote/main');
 const { stat } = require('fs/promises');
+const ffmpeg = require('./ffmpeg');
 
 const logger = require('./logger');
 const menu = require('./menu');
@@ -253,6 +254,8 @@ const readyPromise = app.whenReady();
 
 (async () => {
   try {
+    await ffmpeg.whenImported;
+
     logger.info('Initializing config store');
     await configStore.init();
